Facilitate ARM64 cross-compilation
First, we manually specify whether to use the cmake root path for
find_library. This is to make sure that we actually search the host
system instead of the target sysroot when searching for Vitis libraries.
Second, we extend the platform check to not fail when compiling for
ARM64 systems.